enhance requesting and use new `api` in cli.py
