+++ papercli/__init__.py
@@ -1,4 +1,7 @@
 """
 init python-files
 """
-__all__ = ['papercli', 'api', 'save', 'textgen']
+from papercli.cli import *
+from papercli.save import *
+from papercli.textgen import *
+from papercli.api import *
