+++ papercli/__main__.py
@@ -1,5 +1,5 @@
 """Import cli_main"""
-from papercli.papercli import cli_main
+from cli import cli_main
 
 if __name__ == '__main__':
     cli_main()
